backend/requirements.txt
# Core FastAPI and ASGI server
fastapi[standard]>=0.104.0
uvicorn[standard]>=0.24.0

# Environment configuration
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Authentication and Security
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# QR Code Generation
qrcode[pil]>=7.4.2
Pillow>=10.0.0

# Email validation (Pydantic dependency)
email-validator>=2.0.0
